Create directory for in-progress FWE content (#6636)
This prepares a `/get-started` directory that hides its child pages from
the sitemap and asks search engines to not index them. In-progress FWE
content can be placed here. Eventually custom tutorial functionality and
styles will be applied within the directory as well.

The content is a placeholder so there's something clarifying there, feel
free to change it as needed in other PRs.

**Staged:**
https://dart-dev--pr6636-feat-hidden-fwe-directory-m2cikzvz.web.app/get-started