Clarify rectangle property calculation (#6827)
https://github.com/dart-lang/site-www/issues/6798 reported the
calculation as wrong, but I believe it's right since we tend to document
screen-space where `(0, 0)` is in the upper left and y grows down. To
avoid any future misunderstandings, I've added a doc comment to the
class to clarify this. I've also highlighted the import lines of the
code snippet.

Closes https://github.com/dart-lang/site-www/issues/6798

---------

Co-authored-by: Connie Ooi <connieooi@google.com>