Start migrating the site infrastructure to Jaspr Content (#6822)
Migrates the site infrastructure from NodeJS, pnpm, and 11ty to a
Dart-based setup using [Jaspr](https://jaspr.site) with [Jaspr
Content](https://docs.jaspr.site/content). The three primary motivators
for this change include:

- Simplify site development, contribution, and maintenance, only
requiring Dart.
- Enable us to more easily build and integrate interactive components as
we begin to add more the site.
- Support and improve Jaspr, while also showing off Dart and Jaspr's
capabilities for web and web-adjacent development.

This migration doesn't fully migrate all interactive components and
still maintains some of our existing way of doing things to avoid
impacting the workflow of technical writers too much. Further migrations
and simplification is planned to be completed as part of follow-up work.

This migration provided an amazing chance to test out Jaspr,
particularly its new content-based site support, and help upstream fixes
and improvements. Multiple fixes and improvements have landed to both
Jaspr and Jaspr Content as part of this work and many more developer
experience improvements are planned based on this experience.

--- 

Some of this work, and many of the upstream fixes, were completed by or
with the help of @schultek. Massive thanks to him for Jaspr and all of
his help getting here.

If you're interested, some phases of the migration and incremental fixes
can be found in the previous
https://github.com/dart-lang/site-www/pull/6702 pull request.

Resolves https://github.com/dart-lang/site-www/issues/6824
Resolves https://github.com/dart-lang/site-www/issues/6256
Resolves https://github.com/dart-lang/site-www/issues/5049