Automatically generate hash of generated css file (#6950)
Automatically computes a hash for the generated `main.css` file and
makes it accessible from Dart code. Then appends it as a query parameter
to the `main.css` link from each page. This allows changes to the CSS to
indicate to clients to pull the updated version rather than the
potentially outdated one from cache.

Resolves https://github.com/dart-lang/site-www/issues/6946