Compile client code with dart2wasm (#6953)
This configures `build_web_compilers` to use ddc on dev builds
(`dash_site serve`) while compiling with both `dart2js` and `dart2wasm`
on release modes (`dash_site build`, or the added `--release` flag to
`serve` to experiment with this more easily).

Using jaspr CLI options to switch to dart2wasm would _only_ compile
sources with `dart2wasm`, this uses both compilers and a feature
detection snippet emitted by `dart2wasm` to determine which variant is
used. That seems to be in spirit of
https://github.com/dart-lang/site-www/issues/6854.

To stop jaspr from overriding options passed to `build_web_compilers`,
one needs to use the `--no-managed-build-options` flag. A downside is
that the `PRODUCTION` define set on release modes would not be forwarded
to client code any longer. The constant is only used in the `robots.txt`
generator so this should be fine, forwarding this to client code would
require more changes to `jaspr_cli`.

Closes https://github.com/dart-lang/site-www/issues/6854.