Add custom Jaspr Content layout for FWE (#6977)
Add a custom `learn` layout for FWE that we can use for targeting
customizations and new functionality. Currently it just hides the
sidenav and TOC to give us a blank slate.

This PR also extract the ability to hide the sidenav in wide layouts to
a shared class that different layouts and pages can use.

---

Co-authored-by: Eric Windmill <eric@ericwindmill.com>