Enable secondary markdown output (#6985)
Starts outputting a Markdown version of some routes (the Markdown based
ones) on the website, which can be consumed by tools like LLMs for
including as context with reduces token counts. The implementation is
pulled from docs.flutter.dev which already has this functionality
enabled :)

As we improve and expand this functionality, we'll likely eventually
link to these from near the page headers.

Example for dart.dev/language:
[/language/index.html.md](https://dart-dev--pr6985-misc-secondary-markdown-output-465gt63r.web.app/language/index.html.md)

Resolves https://github.com/dart-lang/site-www/issues/6939