Fix theme flicker on page load (#7007)
In some chromium based browsers, and mobile browsers, the theme flicker
happens if the device theme is set to dark and landing on the website
has 'automatic' theme setting.

**Before**
The flickering starts after 7s in the video.

https://github.com/user-attachments/assets/b2e1a2ec-e610-42de-9ab2-6f3c0a2123a4

**About the Changes:**
Moved the script that already handles the flash case before the main
page contents is loaded.

**After**

https://github.com/user-attachments/assets/484ed6d1-5822-4fc8-ad2d-4beb3dd45ce8