Add initial support for multiple sidenavs (#7006)
Adds support for multiple sidenavs with additional sidenavs being
defined in the `/src/_data/sidenav` folder. Each page can specify a
sidenav by adding the name of a sidenav file to its frontmatter, such as
`sidenav: getStarted`.

Then uses this new functionality to add a custom sidenav for FWE. The
styles for that sidenav be improved as part of follow-up work, but this
makes the underlying functionality work.

This PR also updates the sidenav logic to match that of
docs.flutter.dev, adding support for hidden children.